requirements.txt
celery==4.4.7
Django==3.0.9
djangorestframework==3.11.0
gunicorn==20.0.4
Markdown==3.2.2
django-filter==2.2.0
django-tenants==3.1.0
django-import-export==2.3.0
django-ordered-model==3.4.1
django-colorfield==0.3.2
django-admin-sortable2==0.7.7
django-mailer==2.0
django-solo==1.1.3
django-cors-headers==3.2.1
drf-yasg==1.17.1
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
psycopg2-binary==2.8.5
requests==2.23.0
sentry-sdk==0.14.1
easy-thumbnails==2.7
xhtml2pdf==0.2.5